issue-#1: sample unified apt servlet fixes, some refactoring
